refactor read txns
