--- atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/ReadOnlyTransaction.java
@@ -1,141 +1,0 @@
-/**
- * Copyright 2015 Palantir Technologies
- *
- * Licensed under the BSD-3 License (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://opensource.org/licenses/BSD-3-Clause
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.palantir.atlasdb.transaction.impl;
-
-import java.util.List;
-
-import org.apache.commons.lang3.Validate;
-
-import com.google.common.base.Functions;
-import com.google.common.base.Supplier;
-import com.palantir.atlasdb.cleaner.Cleaner;
-import com.palantir.atlasdb.keyvalue.api.KeyValueService;
-import com.palantir.atlasdb.keyvalue.api.RangeRequest;
-import com.palantir.atlasdb.keyvalue.api.RowResult;
-import com.palantir.atlasdb.keyvalue.api.TableReference;
-import com.palantir.atlasdb.keyvalue.api.Value;
-import com.palantir.atlasdb.transaction.api.AtlasDbConstraintCheckingMode;
-import com.palantir.atlasdb.transaction.api.TransactionReadSentinelBehavior;
-import com.palantir.atlasdb.transaction.service.TransactionService;
-import com.palantir.common.base.AbortingVisitor;
-import com.palantir.common.base.AbstractBatchingVisitable;
-import com.palantir.common.base.BatchingVisitable;
-import com.palantir.common.base.BatchingVisitableFromIterable;
-import com.palantir.common.base.BatchingVisitables;
-import com.palantir.common.base.ClosableIterator;
-import com.palantir.lock.LockRefreshToken;
-import com.palantir.lock.RemoteLockService;
-import com.palantir.timestamp.TimestampService;
-
-/**
- * This will read the values of all committed transactions.
- */
-public class ReadOnlyTransaction extends SnapshotTransaction {
-
-    public ReadOnlyTransaction(KeyValueService keyValueService,
-                               TransactionService transactionService,
-                               long startTimeStamp,
-                               AtlasDbConstraintCheckingMode constraintCheckingMode,
-                               TransactionReadSentinelBehavior readSentinelBehavior,
-                               boolean allowHiddenTableAccess) {
-        super(keyValueService,
-              transactionService,
-              null,
-              startTimeStamp,
-              constraintCheckingMode,
-              readSentinelBehavior,
-              allowHiddenTableAccess);
-    }
-    /* package */ ReadOnlyTransaction(
-            KeyValueService keyValueService,
-            RemoteLockService lockService,
-            TimestampService timestampService,
-            TransactionService transactionService,
-            Cleaner cleaner,
-            Supplier<Long> startTimeStamp,
-            ConflictDetectionManager conflictDetectionManager,
-            SweepStrategyManager sweepStrategyManager,
-            long immutableTimestamp,
-            Iterable<LockRefreshToken> tokensValidForCommit,
-            AtlasDbConstraintCheckingMode constraintCheckingMode,
-            Long transactionTimeoutMillis,
-            TransactionReadSentinelBehavior readSentinelBehavior,
-            boolean allowHiddenTableAccess) {
-        super(keyValueService,
-                lockService,
-                timestampService,
-                transactionService,
-                cleaner,
-                startTimeStamp,
-                conflictDetectionManager,
-                sweepStrategyManager,
-                immutableTimestamp,
-                tokensValidForCommit,
-                constraintCheckingMode,
-                transactionTimeoutMillis,
-                readSentinelBehavior,
-                allowHiddenTableAccess);
-    }
-
-    @Override
-    protected boolean shouldDeleteAndRollback() {
-        // We don't want to delete any data or roll back any transactions because we don't participate in the
-        // transaction protocol.  We just want to skip over anything we find that isn't committed
-        return false;
-    }
-
-    public BatchingVisitable<RowResult<Value>> getRangeWithTimestamps(TableReference tableRef, RangeRequest range) {
-        checkGetPreconditions(tableRef);
-        if (range.isEmptyRange()) {
-            return BatchingVisitables.emptyBatchingVisitable();
-        }
-        return new AbstractBatchingVisitable<RowResult<Value>>() {
-            @Override
-            public <K extends Exception> void batchAcceptSizeHint(
-                    int userRequestedSize,
-                    ConsistentVisitor<RowResult<Value>, K> visitor)
-                    throws K {
-                int requestSize = range.getBatchHint() != null ? range.getBatchHint() : userRequestedSize;
-                int preFilterBatchSize = getRequestHintToKvStore(requestSize);
-
-                Validate.isTrue(!range.isReverse(), "we currently do not support reverse ranges");
-                getBatchingVisitableFromIterator(
-                        tableRef,
-                        range,
-                        requestSize,
-                        visitor,
-                        preFilterBatchSize);
-            }
-
-        };
-    }
-
-    private <K extends Exception> boolean getBatchingVisitableFromIterator(
-            TableReference tableRef,
-            RangeRequest range,
-            int userRequestedSize,
-            AbortingVisitor<List<RowResult<Value>>, K> visitor,
-            int preFilterBatchSize) throws K {
-        ClosableIterator<RowResult<Value>> postFilterIterator =
-                postFilterIterator(tableRef, range, preFilterBatchSize, Functions.identity());
-        try {
-            return BatchingVisitableFromIterable.create(postFilterIterator).batchAccept(userRequestedSize, visitor);
-        } finally {
-            postFilterIterator.close();
-        }
-    }
-
-}
